+++ lib/file.c
@@ -109,6 +109,11 @@ void rm_file_build_path(RmFile *file, char *buf) {
     rm_trie_build_path(&file->session->cfg->file_trie, file->folder, buf, PATH_MAX);
 }
 
+void rm_file_build_dir_path(RmFile *file, char *buf) {
+    g_assert(file);
+    rm_trie_build_path(&file->session->cfg->file_trie, file->folder->parent, buf, PATH_MAX);
+}
+
 RmFile *rm_file_copy(RmFile *file) {
     g_assert(file);