+++ lib/traverse.c
@@ -34,6 +34,7 @@
 
 #include "file.h"
 #include "formats.h"
+#include "logger.h"
 #include "md-scheduler.h"
 #include "preprocess.h"
 #include "utilities.h"