+++ src/rmlint.c
@@ -26,42 +26,18 @@
 #include <locale.h>
 #include <stdlib.h>
 #include <string.h>
+#include<assert.h>
 
 #include "../lib/api.h"
 #include "../lib/config.h"
+#include "../lib/gui.h"
+#include "../lib/logger.h"
+#include "../lib/hash-utility.h"
+#include "../lib/reflink.h"
 
 #if !GLIB_CHECK_VERSION(2, 36, 0)
 #include <glib-object.h>
 #endif
-
-static char *remove_color_escapes(char *message) {
-    char *dst = message;
-    for(char *src = message; src && *src; src++) {
-        if(*src == '\x1b') {
-            src = strchr(src, 'm');
-        } else {
-            *dst++ = *src;
-        }
-    }
-
-    if(dst) {
-        *dst = 0;
-    }
-    return message;
-}
-
-static void logging_callback(_UNUSED const gchar *log_domain,
-                             GLogLevelFlags log_level,
-                             const gchar *message,
-                             gpointer user_data) {
-    RmSession *session = user_data;
-    if(session->cfg->verbosity >= log_level) {
-        if(!session->cfg->with_stderr_color) {
-            message = remove_color_escapes((char *)message);
-        }
-        fputs(message, stderr);
-    }
-}
 
 static void signal_handler(int signum) {
     switch(signum) {
@@ -101,18 +77,42 @@ static void i18n_init(void) {
 #endif
 }
 
+static void maybe_run_alt_main(int argc, const char **argv, char *match_first,
+                               char *alt_main_name, int (*alt_main)(int, const char **)) {
+    if(argc < 2) {
+        return;
+    }
+    if(g_strcmp0(match_first, argv[1]) == 0) {
+        argv[1] = alt_main_name;
+        exit(alt_main(argc - 1, &argv[1]));
+    }
+    for(int i = 2; i < argc; i++) {
+        if(g_strcmp0(match_first, argv[i]) == 0) {
+            rm_log_error_line("%s must be first argument", match_first);
+            exit(EXIT_FAILURE);
+        }
+    }
+}
+
 int main(int argc, const char **argv) {
+#if !GLIB_CHECK_VERSION(2, 36, 0)
+    /* Very old glib. Debian, Im looking at you. */
+    g_type_init();
+#endif
+
     int exit_state = EXIT_FAILURE;
+
     RM_LOG_INIT;
-
-    RmCfg cfg;
-    rm_cfg_set_default(&cfg);
-
-    RmSession session;
-    rm_session_init(&session, &cfg);
-
     /* call logging_callback on every message */
-    g_log_set_default_handler(logging_callback, &session);
+    g_log_set_default_handler(rm_logger_callback, NULL);
+
+    maybe_run_alt_main(argc, argv, "--gui", "shredder", &rm_gui_launch);
+
+    maybe_run_alt_main(argc, argv, "--hash", "rmlint --hash", &rm_hasher_main);
+
+    maybe_run_alt_main(argc, argv, "--is-reflink", "rmlint --is-reflink", &rm_is_reflink_main);
+
+    maybe_run_alt_main(argc, argv, "--dedupe", "rmlint --dedupe", &rm_dedupe_main);
 
     i18n_init();
 
@@ -126,21 +126,17 @@ int main(int argc, const char **argv) {
     sigaction(SIGFPE, &sa, NULL);
     sigaction(SIGABRT, &sa, NULL);
 
-#if !GLIB_CHECK_VERSION(2, 36, 0)
-    /* Very old glib. Debian, Im looking at you. */
-    g_type_init();
-#endif
+
+    RmCfg cfg;
+    rm_cfg_set_default(&cfg);
+    RmSession session;
+    rm_session_init(&session, &cfg);
+
 
     /* Parse commandline */
     if(rm_cmd_parse_args(argc, (char **)argv, &session) != 0) {
         /* Do all the real work */
-        if(cfg.dedupe) {
-            exit_state = rm_session_dedupe_main(&cfg);
-        } else if(cfg.is_reflink) {
-            exit_state = rm_session_is_reflink_main(&cfg);
-        } else {
-            exit_state = rm_cmd_main(&session);
-        }
+        exit_state = rm_cmd_main(&session);
     }
 
     rm_session_clear(&session);